Web Site
A website system (for a client, still in progress). Made with HTML, PHP, CSS, MySQL and JavaScript.
A financial network system that recruit people to join and refer people to join (and get commission for it). Members of the network are exposed to various benefits which vary depending on "Levels"